EPOCH_ENABLE_PROFILER CMake option
